+++ packages/block-logs-stream/src/blockRangeToLogs.ts
@@ -1,29 +1,13 @@
 import { EMPTY, OperatorFunction, concatMap, from, pipe, tap } from "rxjs";
-import { FetchLogsResult, fetchLogs } from "./fetchLogs";
+import { FetchLogsOptions, FetchLogsResult, fetchLogs } from "./fetchLogs";
 import { AbiEvent } from "abitype";
-import { Address, BlockNumber, Client } from "viem";
+import { BlockNumber, UnionOmit } from "viem";
 import { debug } from "./debug";
 
-export type BlockRangeToLogsOptions<abiEvents extends readonly AbiEvent[]> = {
-  /**
-   * [viem `Client`][0] used for fetching logs from the RPC.
-   *
-   * [0]: https://viem.sh/docs/clients/public.html
-   */
-  publicClient: Client;
-  /**
-   * Optional contract address(es) to fetch logs for.
-   */
-  address?: Address | Address[];
-  /**
-   * Events to fetch logs for.
-   */
-  events: abiEvents;
-  /**
-   * Optional maximum block range, if your RPC limits the amount of blocks fetched at a time.
-   */
-  maxBlockRange?: bigint;
-};
+export type BlockRangeToLogsOptions<abiEvents extends readonly AbiEvent[]> = UnionOmit<
+  FetchLogsOptions<abiEvents>,
+  "fromBlock" | "toBlock"
+>;
 
 export type BlockRangeToLogsResult<abiEvents extends readonly AbiEvent[]> = OperatorFunction<
   { startBlock: BlockNumber; endBlock: BlockNumber },
@@ -38,12 +22,9 @@ export type BlockRangeToLogsResult<abiEvents extends readonly AbiEvent[]> = Oper
  * @param {BlockRangeToLogsOptions<AbiEvent[]>} options See `BlockRangeToLogsOptions`.
  * @returns {BlockRangeToLogsResult<AbiEvent[]>} An operator function that transforms a stream of block ranges into a stream of fetched logs.
  */
-export function blockRangeToLogs<abiEvents extends readonly AbiEvent[]>({
-  publicClient,
-  address,
-  events,
-  maxBlockRange,
-}: BlockRangeToLogsOptions<abiEvents>): BlockRangeToLogsResult<abiEvents> {
+export function blockRangeToLogs<abiEvents extends readonly AbiEvent[]>(
+  opts: BlockRangeToLogsOptions<abiEvents>,
+): BlockRangeToLogsResult<abiEvents> {
   let fromBlock: bigint;
   let toBlock: bigint;
 
@@ -57,16 +38,7 @@ export function blockRangeToLogs<abiEvents extends readonly AbiEvent[]>({
     concatMap(() => {
       if (fromBlock > toBlock) return EMPTY;
       debug(`fetching logs for block range ${fromBlock}-${toBlock}`);
-      return from(
-        fetchLogs({
-          publicClient,
-          address,
-          events,
-          fromBlock,
-          toBlock,
-          maxBlockRange,
-        }),
-      ).pipe(
+      return from(fetchLogs<abiEvents>({ ...opts, fromBlock, toBlock })).pipe(
         tap(({ toBlock }) => {
           fromBlock = toBlock + 1n;
         }),