+++ packages/explorer/src/app/(explorer)/[chainName]/worlds/[worldAddress]/observe/useTransactionWatcher.ts
@@ -22,8 +22,8 @@ import { useChain } from "../../../../hooks/useChain";
 import { useWorldAbiQuery } from "../../../../queries/useWorldAbiQuery";
 
 export type WatchedTransaction = {
+  writeId: string;
   hash?: Hex;
-  writeId?: string;
   from?: Address;
   timestamp?: bigint;
   transaction?: Transaction;
@@ -65,9 +65,11 @@ export function useTransactionWatcher() {
         functionName = transaction.input.length > 10 ? transaction.input.slice(0, 10) : "unknown";
       }
 
+      const write = Object.values(observerWrites).find((write) => write.hash === hash);
       setTransactions((prevTransactions) => [
         {
           hash,
+          writeId: write?.writeId ?? hash,
           from: transaction.from,
           timestamp,
           transaction,
@@ -127,7 +129,7 @@ export function useTransactionWatcher() {
         ),
       );
     },
-    [abi, wagmiConfig, worldAddress],
+    [abi, observerWrites, wagmiConfig, worldAddress],
   );
 
   useEffect(() => {
@@ -161,6 +163,8 @@ export function useTransactionWatcher() {
       const writeResult = write.events.find((event): event is Message<"write:result"> => event.type === "write:result");
 
       mergedMap.set(write.hash || write.writeId, {
+        hash: write.hash,
+        writeId: write.writeId,
         from: write.from,
         status: writeResult?.status === "rejected" ? "rejected" : "pending",
         timestamp: BigInt(write.time) / 1000n,