+++ packages/explorer/src/app/(explorer)/layout.tsx
@@ -17,8 +17,8 @@ const jetbrains = JetBrains_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "World Explorer",
-  description: "World Explorer is a tool for visually exploring and manipulating the state of worlds",
+  title: "Worlds Explorer",
+  description: "Worlds Explorer is a tool for visually exploring and manipulating the state of worlds",
 };
 
 export default function RootLayout({