+++ packages/store/ts/storeEventsAbi.ts
@@ -2,6 +2,9 @@ import { parseAbi, AbiEvent } from "abitype";
 import { storeEvents } from "./storeEvents";
 
 export const storeEventsAbi = parseAbi(storeEvents) satisfies readonly AbiEvent[];
+export type storeEventsAbi = typeof storeEventsAbi;
 
 export type StoreEventsAbi = typeof storeEventsAbi;
 export type StoreEventsAbiItem = (typeof storeEventsAbi)[number];
+
+export type StoreSetRecordEventAbiItem = StoreEventsAbiItem & { readonly name: "Store_SetRecord" };