+++ packages/recs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.2.11
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.11
+- @latticexyz/utils@2.2.11
+
 ## 2.2.10
 
 ### Patch Changes