+++ packages/stash/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @latticexyz/stash
 
+## 2.2.11
+
+### Patch Changes
+
+- Updated dependencies [7ddcf64]
+- Updated dependencies [13e5689]
+  - @latticexyz/store@2.2.11
+  - @latticexyz/config@2.2.11
+  - @latticexyz/protocol-parser@2.2.11
+  - @latticexyz/schema-type@2.2.11
+
 ## 2.2.10
 
 ### Patch Changes