+++ packages/world-module-metadata/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @latticexyz/world-module-metadata
 
+## 2.2.11
+
+### Patch Changes
+
+- Updated dependencies [7ddcf64]
+- Updated dependencies [13e5689]
+  - @latticexyz/store@2.2.11
+  - @latticexyz/world@2.2.11
+  - @latticexyz/schema-type@2.2.11
+
 ## 2.2.10
 
 ### Patch Changes